uId类型修改

+++ common/msg/glmsg.go
@@ -2,7 +2,7 @@ package msg
 
 //逻辑服和大厅服消息
 type PlayerBrief struct {
-	UID     int
+	UID     int64
 	Name    string
 	HeadUrl string
 	Icon    int
@@ -10,21 +10,21 @@ type PlayerBrief struct {
 
 //房间简要信息
 type RoomInfo struct {
-	ServerID   int                  //第二类型
-	KindID     int                  //第一类型
-	RoomID     int                  //6位房号
-	NodeID     int                  //在哪个节点上
-	CurCnt     int                  //当前人数
-	MaxCnt     int                  //最多多人数
-	PayCnt     int                  //可玩局数
-	PayType    int                  //支付类型
-	CurPayCnt  int                  //已玩局数
-	CreateTime int64                //创建时间
-	Idx        int                  //服务器标识用的字段
-	IsPublic   bool                 //是否公开匹配
-	Players    map[int]*PlayerBrief //玩家id
-	SvrHost    string               //哪个ip上的房间
-	MachPlayer map[int]struct{}
+	ServerID   int                    //第二类型
+	KindID     int                    //第一类型
+	RoomID     int                    //6位房号
+	NodeID     int                    //在哪个节点上
+	CurCnt     int                    //当前人数
+	MaxCnt     int                    //最多多人数
+	PayCnt     int                    //可玩局数
+	PayType    int                    //支付类型
+	CurPayCnt  int                    //已玩局数
+	CreateTime int64                  //创建时间
+	Idx        int                    //服务器标识用的字段
+	IsPublic   bool                   //是否公开匹配
+	Players    map[int64]*PlayerBrief //玩家id
+	SvrHost    string                 //哪个ip上的房间
+	MachPlayer map[int64]struct{}
 }
 
 ///通知大厅房间结束
